Remove cueing restrictions
The master code restricts cueing near the border, whether or not the
cue actually interferes with the table. This results in a frustrating
playing experience.

For example, if the player attempts to address the bottom of the cue
ball with the cue parallel to the border and five centimetres from it,
the crosshair goes only partway before springing back to centre. The
restriction becomes more prohibitive as the player rotates the view
over the border. The only way to address the bottom of the ball is to
rotate the view over the opposite border (which is an unwelcome
distraction from the ordinary business of directing the cue ball).

Dodge the restrictive code, leaving restrictions to the player’s own
conscience.